+++ src/MatchedBetMate.iOS.Business/Strategy/FreeBetSnrCalculatorStrategy.cs
@@ -11,7 +11,7 @@ namespace MatchedBetMate.iOS.Business.Strategy
     {
         public BetType BetType => BetType.FreeBetSnr;
 
-        public BetCalculationModel Calculate(double backStake, double backOdds, double layOdds, double layCommission)
+        public BetCalculationModel Calculate(decimal backStake, decimal backOdds, decimal layOdds, decimal layCommission)
         {
             var optimalLayStake = (backOdds - 1) / (layOdds - layCommission) * backStake;
